requirements.txt
<<<<<<< HEAD
Django==2.2.16
=======
Django==3.2.16
>>>>>>> 5596f616f928223a336f7372d31359f311bf4a2c
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
<<<<<<< HEAD
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0
django-filter==21.1
=======
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
>>>>>>> 5596f616f928223a336f7372d31359f311bf4a2c
